services/web/requirements.txt
beautifulsoup4==4.9.2
black==20.8b1
colorama==0.4.3
Flask==1.1.2
gunicorn==20.0.4
pre-commit==2.7.1
psycopg2-binary==2.8.4
pytest==6.1.0
